O_Auth/requirements.txt
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauth==1.0.1
python-dotenv==1.2.1
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.38.0
